Move authenticated client http request into auth
